backend/requirements.txt
# Python 3.11.2
Django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
psycopg2===2.9.5
python-dotenv==1.0.0
djoser==2.1.0
django-filter==22.1
django-cors-headers==3.14.0
Pillow==9.4.0
